+++ class/actor.lua
@@ -112,12 +112,12 @@ methods_actor = {
     end,
 
 
-    fire_bullet = function(self, x, y, range, direction, damage, pierce_multiplier, hit_sprite, tracer)
+    fire_bullet = function(self, x, y, range, direction, damage, pierce_multiplier, hit_sprite, tracer, no_proc)
         -- Set whether or not the bullet damager can pierce
         local can_pierce = false
         if pierce_multiplier then can_pierce = true end
 
-        local inst = GM._mod_attack_fire_bullet(self.value, x, y, range, direction, damage, hit_sprite or -1, can_pierce, true)
+        local inst = GM._mod_attack_fire_bullet(self.value, x, y, range, direction, damage, hit_sprite or -1, can_pierce, not no_proc)
         local damager = inst.attack_info
         damager.instance = Wrap.unwrap(inst)
         damager.damage_color = Color.WHITE_ALMOST
@@ -136,8 +136,8 @@ methods_actor = {
     end,
 
 
-    fire_explosion = function(self, x, y, width, height, damage, explosion_sprite, sparks_sprite)
-        local inst = GM._mod_attack_fire_explosion(self.value, x, y, width, height, damage, explosion_sprite or -1, sparks_sprite or -1, true)
+    fire_explosion = function(self, x, y, width, height, damage, explosion_sprite, sparks_sprite, no_proc)
+        local inst = GM._mod_attack_fire_explosion(self.value, x, y, width, height, damage, explosion_sprite or -1, sparks_sprite or -1, not no_proc)
         local damager = inst.attack_info
         damager.instance = Wrap.unwrap(inst)
         damager.damage_color = Color.WHITE_ALMOST
@@ -146,7 +146,7 @@ methods_actor = {
     end,
 
 
-    fire_explosion_local = function(self, x, y, width, height, damage, explosion_sprite, sparks_sprite)
+    fire_explosion_local = function(self, x, y, width, height, damage, explosion_sprite, sparks_sprite, no_proc)
         local mask = gm.constants.sBite1Mask
         self.value:fire_explosion_local(0, x, y, damage, sparks_sprite or -1, 2, width / GM.sprite_get_width(mask), height / GM.sprite_get_height(mask))
         local inst = gm.variable_global_get("attack_bullet")
@@ -163,10 +163,10 @@ methods_actor = {
     end,
 
 
-    fire_direct = function(self, target, damage, direction, x, y, hit_sprite)
+    fire_direct = function(self, target, damage, direction, x, y, hit_sprite, no_proc)
         target = Wrap.unwrap(target)
 
-        local inst = GM._mod_attack_fire_direct(self.value, target, x or target.x, y or target.y, direction or 0, damage, hit_sprite or -1, true)
+        local inst = GM._mod_attack_fire_direct(self.value, target, x or target.x, y or target.y, direction or 0, damage, hit_sprite or -1, not no_proc)
         local damager = inst.attack_info
         damager.instance = Wrap.unwrap(inst)
         damager.damage_color = Color.WHITE_ALMOST